+++ tools/onnx-graphsurgeon/tests/test_importers.py
@@ -18,7 +18,7 @@ from onnx_graphsurgeon.importers.onnx_importer import OnnxImporter
 from onnx_graphsurgeon.logger.logger import G_LOGGER
 from onnx_graphsurgeon.ir.tensor import Tensor, Variable, Constant
 
-from onnx_models import identity_model, lstm_model, scan_model, dim_param_model
+from onnx_models import identity_model, lstm_model, scan_model, dim_param_model, initializer_is_output_model
 
 from collections import OrderedDict
 import onnx.shape_inference
@@ -86,7 +86,7 @@ class TestOnnxImporter(object):
         assert node.attrs["strings_attr"] == strings_attr
 
 
-    @pytest.mark.parametrize("model", [identity_model(), lstm_model(), scan_model()])
+    @pytest.mark.parametrize("model", [identity_model(), lstm_model(), scan_model(), dim_param_model(), initializer_is_output_model()])
     def test_import_graph(self, model):
         graph = OnnxImporter.import_graph(model.load().graph)
         model.assert_equal(graph)