+++ tools/onnx-graphsurgeon/CHANGELOG.md
@@ -1,12 +1,12 @@
-# ONNX GraphSurgeon changelog history
+# ONNX GraphSurgeon Change Log
 
 Dates are in YYYY-MM-DD format.
 
-## vNext
-
+## v0.2.4
 ### Fixed
 - `fold_constants()` will no longer fail if there is nothing to fold in the graph
 - `cleanup()` will now properly remove the producer nodes of graph inputs.
+- Fixed a bug where graph input/output tensors not attached to nodes would not be correctly exported.
 
 
 ## v0.2.3